% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_cdf.R
\name{vsk_cdf}
\alias{vsk_cdf}
\title{Calculating the cumulative distribution function of Vasicek}
\usage{
vsk_cdf(x, Rho, P)
}
\arguments{
\item{x}{A numeric vector in the [0, 1] interval that is supposed to
follow the Vasicek distribution}

\item{Rho}{The Rho parameter in the Vasicek distribution}

\item{P}{The P parameter in the Vasicek distribution}
}
\value{
A numeric vector with the corresponding cdf.
}
\description{
The function \code{vsk_cdf} calculates the cumulative distribution 
function of Vasicek.
}
\examples{
vsk_cdf(c(0.278837772815679, 0.5217229060260343), Rho = 0.2, P = 0.3)
# [1] 0.5 0.9
}
