% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_dmm.R
\name{vsk_dmm}
\alias{vsk_dmm}
\title{Estimating Vasicek parameters by using direct moment matching}
\usage{
vsk_dmm(x)
}
\arguments{
\item{x}{A numeric vector in the (0, 1) interval that is supposed to
follow the Vasicek distribution}
}
\value{
A list with Vasicek parameters, namely Rho and P.
}
\description{
The function \code{vsk_mle} estimates parameters in the Vasicek
distribution by using direct moment matching.
}
\examples{
vsk_dmm(vsk_rvs(1000, Rho = 0.2, P = 0.1))
# $Rho
# [1] 0.2135844
# $P
# [1] 0.1025469
}
