% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_imm.R
\name{vsk_imm}
\alias{vsk_imm}
\title{Estimating Vasicek parameters by using indirect moment matching}
\usage{
vsk_imm(x)
}
\arguments{
\item{x}{A numeric vector in the (0, 1) interval that is supposed to
follow the Vasicek distribution}
}
\value{
A list with Vasicek parameters, namely Rho and P.
}
\description{
The function \code{vsk_imm} estimates parameters in the Vasicek
distribution by using indirect moment matching.
}
\examples{
vsk_imm(vsk_rvs(1000, Rho = 0.2, P = 0.1))
# $Rho
# [1] 0.2110422
# $P
# [1] 0.1024877
}
