% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_mle.R
\name{vsk_mle}
\alias{vsk_mle}
\title{Estimating Vasicek parameters by using maximum likelihood estimator}
\usage{
vsk_mle(x)
}
\arguments{
\item{x}{A numeric vector in the (0, 1) interval that is supposed to
follow the Vasicek distribution}
}
\value{
A list with Vasicek parameters, namely Rho and P.
}
\description{
The function \code{vsk_mle} estimates parameters in the Vasicek
distribution by using maximum likelihood estimator.
}
\examples{
vsk_mle(vsk_rvs(1000, Rho = 0.2, P = 0.1))
# $Rho
# [1] 0.2110976
# $P
# [1] 0.1025469
}
