% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_pdf.R
\name{vsk_pdf}
\alias{vsk_pdf}
\title{Calculating the probability density function of Vasicek}
\usage{
vsk_pdf(x, Rho, P)
}
\arguments{
\item{x}{A numeric vector in the (0, 1) interval that is supposed to
follow the Vasicek distribution}

\item{Rho}{The Rho parameter in the Vasicek distribution}

\item{P}{The P parameter in the Vasicek distribution}
}
\value{
A numeric vector with the corresponding pdf.
}
\description{
The function \code{vsk_pdf} calculates the probability density
function of Vasicek.
}
\examples{
vsk_pdf(c(0.01, 0.02), Rho = 0.2, P = 0.3)
# [1] 0.07019659 0.22207564
}
