% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_ppf.R
\name{vsk_ppf}
\alias{vsk_ppf}
\title{Calculating the percentile point function of Vasicek}
\usage{
vsk_ppf(Alpha, Rho, P)
}
\arguments{
\item{Alpha}{A numeric vector of probabilities}

\item{Rho}{The Rho parameter in the Vasicek distribution}

\item{P}{The P parameter in the Vasicek distribution}
}
\value{
A numeric vector with the corresponding ppf.
}
\description{
The function \code{vsk_ppf} calculates the percentile point
function of Vasicek.
}
\examples{
vsk_ppf(c(0.5, 0.9), Rho = 0.2, P = 0.3)
# [1] 0.2788378 0.5217229
}
