% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsk_qbe.R
\name{vsk_qbe}
\alias{vsk_qbe}
\title{Estimating Vasicek parameters by using quantile-based estimator}
\usage{
vsk_qbe(x)
}
\arguments{
\item{x}{A numeric vector in the (0, 1) interval that is supposed to
follow the Vasicek distribution}
}
\value{
A list with Vasicek parameters, namely Rho and P.
}
\description{
The function \code{vsk_qbe} estimates parameters in the Vasicek
distribution by using quantile-based estimator. It is not recommended
for small sample size.
}
\examples{
vsk_qbe(vsk_rvs(1000, Rho = 0.2, P = 0.1))
# $Rho
# [1] 0.1941091
# $P
# [1] 0.1019701
}
